Safari-611.1.10/LayoutTests/editing/pasteboard/entries-api/datatransfer-items-drop-directoryReader-error-expected.txt

Error case coverage for fileSystemDirectoryReader.readEntries()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dataTransfer.items.length is 1
PASS ex.name is "InvalidStateError"
PASS successfullyParsed is true

TEST COMPLETE
